+++ SAP2000_oM/Fragments/IPanelOffset.cs
@@ -20,12 +20,6 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using BH.oM.Quantities.Attributes;
 using System.ComponentModel;
 using BH.oM.Base;
 
@@ -34,22 +28,5 @@ namespace BH.oM.Adapters.SAP2000.Fragments
     [Description("Base interface for panel offsets in SAP. Contains only the type of offset.")]
     public partial interface IPanelOffset : IFragment
     {
-    }
-
-    public class PanelOffsetByJointPattern : IPanelOffset
-    {
-        [Description("This is the name of the defined joint pattern that is used to calculate the joint offsets.")]
-        public virtual string OffsetPattern { get; set; } = "";
-
-        [Length]
-        [Description("This is the scale factor applied to the joint pattern when calculating the joint offsets.")]
-        public virtual double OffsetPatternSF { get; set; } = 0;
-    }
-
-    public class PanelOffsetByPoint : IPanelOffset
-    {
-        [Length]
-        [Description("This is an array of joint offsets for each of the points that define the area object.")]
-        public virtual double[] Offset { get; set; } = null;
     }
 }
