+++ SAP2000_oM/SAP2000_oM.csproj
@@ -72,12 +72,20 @@
     <Compile Include="Enums\BarDesignProcedureType.cs" />
     <Compile Include="Enums\BarInsertionPointLocation.cs" />
     <Compile Include="Enums\SDShapeType.cs" />
-    <Compile Include="Fragments\PanelOffset.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByCookieCutLines.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByCookieCutPoints.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByGeneralDivide.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByMaximumSize.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByNumberOfObjects.cs" />
+    <Compile Include="Fragments\PanelAutoMeshByPointsOnEdges.cs" />
+    <Compile Include="Fragments\IPanelOffset.cs" />
     <Compile Include="Fragments\PanelEdgeConstraint.cs" />
-    <Compile Include="Fragments\PanelAutoMesh.cs" />
+    <Compile Include="Fragments\IPanelAutoMesh.cs" />
     <Compile Include="Fragments\BarInsertionPoint.cs" />
     <Compile Include="Fragments\BarDesignProcedure.cs" />
     <Compile Include="Fragments\BarAutoMesh.cs" />
+    <Compile Include="Fragments\PanelOffsetByJointPattern.cs" />
+    <Compile Include="Fragments\PanelOffsetByPoint.cs" />
     <Compile Include="Fragments\SAP2000Id.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Enums\ShellType.cs" />
